+++ rust/src/common.rs
@@ -1,0 +1,15 @@
+pub struct Unset;
+pub struct Get;
+pub struct Post;
+
+impl Get {
+    pub const fn name() -> &'static str {
+        "GET"
+    }
+}
+
+impl Post {
+    pub const fn name() -> &'static str {
+        "POST"
+    }
+}
